uns_mcp_nats/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uns-mcp-nats"
version = "0.1.0"
authors = [
    {name = "Dronomy.io", email = "info@dronomy.io"},
]
description = "UNS-MCP with NATS transport integration"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp>=0.1.0",
    "nats-py>=2.4.0",
    "unstructured-client>=0.13.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "UNS-MCP>=0.1.0",
]

[project.optional-dependencies]
client = [
    "anthropic>=0.6.0",
    "rich>=13.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[project.urls]
"Homepage" = "https://github.com/dronomyio/nats-transport-for-MCP"
"Bug Tracker" = "https://github.com/dronomyio/nats-transport-for-MCP/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]